Results plotting and prev target

+++ lln/exp/ExperimentSummerizer.py
@@ -5,11 +5,13 @@ import json
 import itertools
 import pandas as pd
 import seaborn as sns
+import pickle
 sns.set_style("whitegrid")
 import matplotlib.pyplot as plt
 from lln.plotting.seaborn.plots import highlight_legend_titles
 import matplotlib.font_manager as fm
 from lln.exp.formatted_summary import shape_format_summarized_df
+from lln.exp.results_plotting import per_experiment_confusion_matrix, per_experiment_feature_attributions, per_experiment_per_timepoint_confusion_matrix
 
 class ExperimentSummerizer:
     '''A class that summarizes and extracts certain results from an experiment
@@ -30,6 +32,19 @@ class ExperimentSummerizer:
         for run_name in self.run_names:
             self.summarize_run(run_name, state_selection_dataset, state_selection_metric, higher_is_better)
         self.average_runs()
+        self.plot_per_experiment_results(labels=self.config['labels'], subject_splits=['Train', 'Val', 'Test'], feature_attributions=False)
+        
+    def plot_per_experiment_results(self, labels, subject_splits = ['Train', 'Val', 'Test'], feature_attributions=False):
+        '''Plot the results of the experiments as confusion matrices.'''
+        exps_df = merge_in_df(self.exps_path, [self.exp_name], subject_splits=subject_splits)
+        exps_dfs = {split_name: exps_df[exps_df['split'] == split_name] for split_name in subject_splits}
+        for split_name in subject_splits:
+            for weighted in [True, False]:
+                per_experiment_per_timepoint_confusion_matrix(exps_dfs[split_name], self.exps_path, exp_name=self.exp_name, exp_better_name=self.exp_name, labels=labels, seeds=[0], weighted=weighted, split_name=split_name)
+                per_experiment_confusion_matrix(exps_dfs[split_name], self.exps_path, exp_name=self.exp_name, exp_better_name=self.exp_name, labels=labels, seeds=[0], weighted=weighted, split_name=split_name)
+            if feature_attributions:
+                for only_matching in [None, 0, 1]:
+                    per_experiment_feature_attributions(exps_dfs[split_name], self.exps_path, exp_name=self.exp_name, exp_better_name=self.exp_name, labels=labels, seeds=[0], only_matching=only_matching, split_name=split_name)
     
     def summarize_run(self, run_name, state_selection_dataset='Val', state_selection_metric='B-Acc.', higher_is_better=True):
         '''Export a results.csv file for each run, containing the best epoch for each metric and dataset.'''
@@ -62,26 +77,27 @@ class ExperimentSummerizer:
     def average_runs(self):
         '''Summarizes the results of multiple experiment runs, e.g. splits or seeds. Each run should 
         have a subdirectory with the corresponding name inside the experiment directory.'''
-        results_df = None
-        for run_name in self.run_names:
-            run_path = os.path.join(self.exp_path, run_name)
-            assert os.path.exists(run_path), f"Experiment {self.exp_name}, run {run_name} not found"
-            df = pd.read_csv(os.path.join(run_path, f'results{self.config_suffix}.csv'))
-            df.insert(0, 'Run', run_name)
-            results_df = df if results_df is None else pd.concat([results_df, df], ignore_index=True)
-        results_df = self._set_rows_mean_std(results_df)
-        results_df.to_csv(os.path.join(self.exp_path, f'results{self.config_suffix}.csv'), index=False)
-            
-        # Average the loss_trajectory and progress files
-        for file_name in ['loss_trajectory', 'progress']:
-            progress_df = None
+        if not os.path.exists(os.path.join(self.exp_path, f'results{self.config_suffix}.csv')):
+            results_df = None
             for run_name in self.run_names:
                 run_path = os.path.join(self.exp_path, run_name)
-                df = pd.read_csv(os.path.join(run_path, 'trainer', file_name+'.csv'))
+                assert os.path.exists(run_path), f"Experiment {self.exp_name}, run {run_name} not found"
+                df = pd.read_csv(os.path.join(run_path, f'results{self.config_suffix}.csv'))
                 df.insert(0, 'Run', run_name)
-                progress_df = df if progress_df is None else pd.concat([progress_df, df], ignore_index=True)
-            progress_df = self._set_rows_mean_std(progress_df, non_avg=['Dataset', 'Epoch'])
-            progress_df.to_csv(os.path.join(self.exp_path, file_name+'.csv'), index=False)
+                results_df = df if results_df is None else pd.concat([results_df, df], ignore_index=True)
+            results_df = self._set_rows_mean_std(results_df)
+            results_df.to_csv(os.path.join(self.exp_path, f'results{self.config_suffix}.csv'), index=False)
+                
+            # Average the loss_trajectory and progress files
+            for file_name in ['loss_trajectory', 'progress']:
+                progress_df = None
+                for run_name in self.run_names:
+                    run_path = os.path.join(self.exp_path, run_name)
+                    df = pd.read_csv(os.path.join(run_path, 'trainer', file_name+'.csv'))
+                    df.insert(0, 'Run', run_name)
+                    progress_df = df if progress_df is None else pd.concat([progress_df, df], ignore_index=True)
+                progress_df = self._set_rows_mean_std(progress_df, non_avg=['Dataset', 'Epoch'])
+                progress_df.to_csv(os.path.join(self.exp_path, file_name+'.csv'), index=False)
 
     def _set_rows_mean_std(self, df, non_avg = ['Dataset']):
         non_avg_values = [[(x, col) for x in df[col].unique()] for col in non_avg]
@@ -174,3 +190,32 @@ class ExperimentsSummerizer:
                 else:
                     plt.show()
                 plt.close()
+                
+def merge_in_df(exps_path, exp_names, exp_better_names = dict(), subject_splits = ['Val', 'Test']):
+    data = []
+    for exp_name in exp_names:
+        exp_path = os.path.join(exps_path, exp_name)
+        config = json.load(open(os.path.join(exp_path, "config_train.json")))
+        for split in range(config['nr_splits']):
+            for seed in config['seeds']:
+                run_name = f"SPLIT_{split}_SEED_{seed}"
+                run_splits = json.load(open(os.path.join(exp_path, run_name, "split.json"), 'rb'))
+                targets = pickle.load(open(os.path.join(exp_path, run_name, "model_outputs", "targets.pkl"), 'rb'))
+                predictions = pickle.load(open(os.path.join(exp_path, run_name, "model_outputs", "predictions.pkl"), 'rb'))
+                nonpadded_rows = pickle.load(open(os.path.join(exp_path, run_name, "model_outputs", "nonpadded_rows.pkl"), 'rb'))
+                #attributions = pickle.load(open(os.path.join(exp_path, run_name, "model_outputs", "attributions.pkl"), 'rb'))
+                for subject_split in subject_splits:
+                    for subject in run_splits[subject_split]:
+                        for t_ix in range(len(targets[subject])):
+                            if nonpadded_rows[subject][t_ix] > 0:
+                                data.append({
+                                    "subject": subject,
+                                    "t_ix": t_ix,
+                                    "target": targets[subject][t_ix],
+                                    "pred": predictions[subject][t_ix],
+                                    #"attribution": attributions[subject][t_ix],
+                                    "exp": exp_better_names.get(exp_name, exp_name),
+                                    "seed": seed,
+                                    "split": subject_split
+                                })
+    return pd.DataFrame(data)
